Return err if unlock fails
Caught this when testing the step-down function and ensuring that it works.
It seems that about 50% of the time the lock will be returned to the current
node. I think the 5 second sleep may be the reason for this although I would
still expect the stanby node to take over since after step-down the former
leader sleeps for 10 seconds.