[etcd] use clientv3 forcibly if user requested
etcd storage auto detects cluster version.  Unfortunately though,
the auto detection sometimes detect v3 cluster as v2.

Worse, even if a user specifies the cluster version in HCL as

    storage "etcd" {
        etcd_api = "v3"
        ...
    }

vault denies to start with this error message:

    etcd3 is required: etcd2 is running

This commit eliminates this version check to work around the
auto detection problem because the check is in fact unnecessary
as the user explicitly requested to connect with v3 protocol.