Recreate LDAP client object after authenticating
In the flow of binding to ldap, searching for the user, binding again
as that user, and then finally binding back as the original bind user,
create a fresh LDAP client object before the final rebinding step.
Without this, the rebinding step fails when the ldap server is DUO's
ldap proxy.