auth/aws: Add support for bound_cidr
This adds support for specifying a bound_cidr parameter on AWS auth role
entries. The singular name of "bound_cidr" is chosen instead of the
plural "bound_cidrs" as the other bindings in AWS auth roles are all
singularly named but many of them take multiple values as well, so this
maintains that consistency.

Fixes #5946