Move fake test tidy delay to make the race condition test not flakey
Previously this race condition test would fail around 10% of the time on
my own machine. By moving the delay to after listing the secretIDHMACs
I'm able to get it to fail 100% of the time when removing the
`lock.Lock() lock.Unlock()` code that has been added in this pull
request.