Add Docs For Prometheus Metrics
Prometheus metrics were added as part
of the Vault v1.1.0 release in PR #5308.
But no documentation was created.

This adds the configuration docs, but
does not include the missing API docs. It
also looks like the OpenAPI endpoint is missing
the the details for /v1/sys/metrics?format=prometheus
which is also not fixed with this change.