Remove PostgreSQLLock.held: it doesn't add anything except a failure mode.
Add PostgreSQLLock.renewTicker to manage the lock renewal lifecycle.

Add PostgreSQLLock fields ttl, renewInterval, and retryInterval: this
allows modifying the default values in tests.

In Lock(), use the stopCh argument to control goroutine termination
instead of making out own.  Remove unused done channel.

tryToLock/writeItem() weren't doing quite what the dynamodb versions that
inspired them did: in the dynamodb version, when a lock wasn't obtained
because it was held by another instance, the lock attempt would be
retried.  In this version we weren't doing that because any failure to
grab the lock for whatever reason was treated as an error and resulted in
a write to the errors chan.

To address this writeItem now also returns a bool indicating whether it
wrote exactly 1 row, indicating that the upsert succeeded.  When
(false, nil) is returned that means no error occurred but also that no
lock was obtained, so tryToLock will retry.

periodicallyRenewLock() exits when it doesn't lock successfully, and
closes its done channel.  This obviates the need for watch(), which has
been removed.

Unlock() stops the ticker used in periodicallyRenewLock, which
saves us from falling victim to the same problem as #5828.

Added testPostgreSQLLockRenewal which attempts to duplicate in spirit
the test from #6512, and testPostgresSQLLockTTL which is the exact
equivalent of the corresponding DynamoDB test.