sdk/framework: do not call Validate in HandleRequest
Calling Validate in HandleRequest before any downstream handler has been
called prevents a handler from overriding the default schema conversion
rules which can be useful. For example, the default conversion for a
TypeDurationSecond field will return an error if a negative duration is
specified. However, allowing a configuration field to accept positive,
zero, and negative durations allows for the setting, clearing, and
disabling of the configuration.