Add fine-grained locking for file audit sink.
This replaces the exclusive mutex that protected the logging of
requests and responses. This limited the overall throughput when
file auditng is enabled as only a single goroutine at a time could
be marshalling request/response data.

This creates a new mutex which only protects write access to the
the file. The lock is only taken after the data has marshalled and
is ready for writing to the file.