When cleaning up a TestCluster, set each core's logger to level Error. This eliminates a bunch of noise at the end of a test's logs, making it easier to diagnose test failures.
Add TestCluster.Logger.  This is intendend to be used when test helpers
want to log something. If TestClusterOptions.Logger is non-nil, it will
be used as the cluster logger, otherwise we create one based on test
name.  In the absence of CoreConfig.Logger, this logger is also used as
the parent of each core's log.  This makes it easy when running parallel
tests to identify which log message came from which test.