Convert token_test to use the docker ldap server.
Attempt to convert TestIdentityStore_Integ_GroupAliases to use the
docker ldap server, but it's not working yet.  The first obstacle was
that it expects a user to exist in two external group: Scientists and
Italians.  There are no users in the new ldap server's data who belong
to two ldap groups.

One thing Becca was saying she hoped the docker ldap server would allow
us is to make changes to the LDAP data.  So I decided to modify the test
such that it first sees a user in a single LDAP group, then once we add
the user to a second group, upon re-logging in we see them in that
identity group as well.

This proved not to be so easy.  I got

  LDAP Result Code 8 "Strong Auth Required": modifications require authentication

Some googling suggested this would be fixed if we used TLS.  This was
a bit of an adventure since they use a self-signed cert, but eventually
I got that working, at least in the context of a test: I wasn't able to
get the LDAP backend itself to be happy with that, which is why
PrepareTestContainer now returns two ldaputil.ConfigEntry structs.

Anyway, that turned out not to be the right fix, or possible just
insufficient.  I'm still getting the same error, but now I think it's
because the admin user hasn't been granted write privs.  I'm not
certain how to do that without building a new docker image, and if that
is required I might revisit which docker image we use.