Create .bundle and set group when running container
If this is not set, `make website` fails due to permission errors in the docker container
Fixes #5589. 

Side note: it takes **a lot** to build and run the application. I believe several stuff could be improved here, starting with the fact that the `netlify` image is about 4gb which seems completely unnecessary.