don't revoke token when c.expiration.RegisterAuth fails
Looks like when expiration.RegisterAuth fails, we revoke the token
initiating the request rather than a newly created token.

If the underlying storage happens to be unavailable just during the
expiration.RegisterAuth but available for the subsequent revoke orphan
call, the request token is invalidated and operators are forced to
rotate it manually.

This commit illustrates the issue with a unit test that forces
RegisterAuth unavailability.  Without commenting out
`c.expiration.RegisterAuth` lines in vault/request_handling.go , the
test fails with the following:

```
--- FAIL: TestLogical_CreateToken_Failure (0.33s)
    logical_test.go:365: ROOT TOKEN: 1Zfy8CqG8k8n4ok2UwwMpoFi
    logical_test.go:373: NEW TOKEN:  6GkVz5y3JtoUqXB0RcRVQIco
    --- FAIL: TestLogical_CreateToken_Failure/lookup_token_after_failed_request (0.00s)
        http_test.go:113: Expected status 200, got 403. Body:

            {"errors":["bad token"]}
    --- FAIL: TestLogical_CreateToken_Failure/create_another_token_after_failed_request (0.00s)
        http_test.go:113: Expected status 200, got 403. Body:

            {"errors":["permission denied"]}
```