Retry on failed lease collection
On the C* storage backend, collecting all the leases is a really
expensive operation and often times out. This operation is attempted
when a node becomes leader, and the time out leads a node to seal. This
means in practice that a leadership election often leads to one or two
nodes sealing before a lease restore happens to succeed. It seems
harmless to retry collecting the leases a few times before giving up and
sealing.