Enable root user credential rotation in MongoDB
This takes its logic from the SetCredentials function with some changes
(ex: it's generating a password rather than taking one as a parameter).

This will error if the username isn't specified in the config. Since
Mongo defaults to unauthorized, this seemed like an easy check to make
to prevent strange behaviors when it tries to rotate the "" user.

I've also overhauled the runCommandWithRetry function to handle timeouts
better. Particularly context timeouts. This wraps the provided
context with a timeout in case one isn't provided (if one is already
set, the shorter of the two will be the effective one).