Update go.mod to corrected plugin tags
This addresses an issue found in #8696 which was determined to be due tothe Go module proxy having a cached copy of a tag that doesn't match theofficial version (due a build prep error weeks ago). All of the reposgot new patch versions, but the content is identical.