Allow auto_auth with templates without specifying a sink
For situations where you want the Vault agent to handle one or
more templates but do not require the acquired credentials
elsewhere.

Modify the logic in SyncServer so that if there are no sinks,
ignore any new credentials. Since SyncServer is responsible
for shutting down the agent, make sure it still properly
shuts down in this new situation.

Updated documentation.

Solves #7988